requirements.txt
configparser==3.5.0
flake8==3.3.0
pytest==3.1.2
pytest-base-url==1.4.0
pytest-selenium==1.10.0
pytest-variables==1.6.1
pytest-html
pyotp==2.2.6
PyPOM==1.1.1
tox
